django/requirements.txt
python-decouple
django[argon2]
django-allauth
django-nh3
django-redis
djangorestframework
djangorestframework-simplejwt
psycopg2
certbot
certbot-nginx
gunicorn
cryptography
requests